moved basic modpack handler into their own file && reworked ModpackDownload
